+++ templates/shortcode-comments.php
@@ -2,9 +2,6 @@
 
 use \AnyComment\Helpers\AnyCommentTemplate;
 
-/**
- * This template is used to display comments.
- */
 if (!defined('ABSPATH')) {
     exit; // Exit if accessed directly
 }