+++ CHANGELOG.md
@@ -6,10 +6,12 @@
 
 - Changed cache directory directory from `wp-content/uploads/cache/anycomment` to `wp-content/cache/anycomment`
 - Plugin interface is not translated to Ukrainian, more to come  
+- Main plugin character was updated in admin dashboard and for users without avatar
 
 **Fixes:**
 
 - Fixed multiple installation bugs, when it was not possible to upload files or make likes, #411
+- Fixed issue when comments when they were closed
 
 ## 0.0.91 – 08.09.2019